requirements.txt
Django==1.11.1
celery
django-celery-results
django-debug-toolbar
django-recaptcha2
django-rest-framework
django-bootstrap3-datetimepicker
django-timezone-field
django-recurrence
